Move OnionShare object into its own file, so __init__ is just CLI logic
